+++ packages/test-helpers/async_multi.js
@@ -1,6 +1,47 @@
 // This depends on tinytest, so it's a little weird to put it in
 // test-helpers, but it'll do for now.
 
+// Provides the testAsyncMulti helper, which creates an async test
+// (using Tinytest.addAsync) that tracks parallel and sequential
+// asynchronous calls.  Specifically, the two features it provides
+// are:
+// 1) Executing an array of functions sequentially when those functions
+//    contain async calls.
+// 2) Keeping track of when callbacks are outstanding, via "expect".
+//
+// To use, pass an array of functions that take arguments (test, expect).
+// (There is no onComplete callback; completion is determined automatically.)
+// Expect takes a callback closure and wraps it, returning a new callback closure,
+// and making a note that there is a callback oustanding.  Pass this returned closure
+// to async functions as the callback, and the machinery in the wrapper will
+// record the fact that the callback has been called.
+//
+// A second form of expect takes data arguments to test for.
+// Essentially, expect("foo", "bar") is equivalent to:
+// expect(function(arg1, arg2) { test.equal([arg1, arg2], ["foo", "bar"]); }).
+//
+// You cannot "nest" expect or call it from a callback!  Even if you have a chain
+// of callbacks, you need to call expect at the "top level" (synchronously)
+// but the callback you wrap has to be the last/innermost one.  This sometimes
+// leads to some code contortions and should probably be fixed.
+
+// Example: (at top level of test file)
+//
+// testAsyncMulti("test name", [
+//   function(test, expect) {
+//     ... tests here
+//     Meteor.defer(expect(function() {
+//       ... tests here
+//     }));
+//
+//     call_something_async('foo', 'bar', expect('baz')); // implicit callback
+//
+//   },
+//   function(test, expect) {
+//     ... more tests
+//   }
+// ]);
+
 var ExpectationManager = function (test, onComplete) {
   var self = this;